Small tweaks.
Check for Textures not working correctly.  Please ensure the proper textures are in the Themes Folder or that the "Themed" setting is set to "false" in the config file before loading the addon.  Loading the adding with the theming engines enabled, but without the textures will result in a crash.  I am hoping to have this fixed soon.