Added Text Colorization based on HP values.
2/3+ HP:  White Text
1/2-2/3 HP:  Yellow Text
Less than 1/2 HP:  Red Text

This is an approximation of blood agro adjustments.  All text objects for a given player in the party list will adjust based on that player's HP values.  While possible to restrict this to purely the HP text, I think it's more beneficial for all text for that player to reflect this.