Added Exemplar Points and Master Levels into the Player Stats Panel.
*KNOWN ISSUE*
Due to how FFXI updates the player information in-game, the exp bar, current exp value, the merit, JP, and ML equivalents, and the current JP totals may be outdated.  This is due to the fact that FFXI does not update these values until zoning, logging in, or opening the status menu.  If the values lag behind,  opening the status menu will update them.